[add xml view]  rating bar xml 을 compose 에 추가

+++ android/fast_30/android_compose/Portfolio/app/src/main/java/com/example/portfolio/ui/screen/home/HomeViewModel.kt
@@ -80,5 +80,5 @@ data class NearRestaurantInfo(
     val address: String,
     val lon: Double,
     val lat: Double,
-    val rating: Double = Random.nextDouble(0.0, 5.0)
+    val rating: Float = Random.nextDouble(0.0, 5.0).toFloat()
 )
